requirements.txt
langchain==0.0.286
langchain_experimental==0.0.16
faiss-cpu==1.7.4
# openai==0.27.8  
# pdfminer==20191125
# beautifulsoup4==4.12.2
# python-dotenv==1.0.0
# tiktoken==0.4.0
streamlit==1.24.0
google-search-results==2.4.2
